Move intersection models to new package
