+++ src/main/java/org/opentripplanner/routing/core/intersection_model/IntersectionTraversalCostModel.java
@@ -1,5 +1,6 @@
-package org.opentripplanner.routing.core;
+package org.opentripplanner.routing.core.intersection_model;
 
+import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.api.request.RoutingRequest;
 import org.opentripplanner.routing.vertextype.IntersectionVertex;
