+++ src/main/java/org/opentripplanner/routing/core/intersection_model/SimpleIntersectionTraversalCostModel.java
@@ -1,8 +1,9 @@
-package org.opentripplanner.routing.core;
+package org.opentripplanner.routing.core.intersection_model;
 
 import java.io.Serializable;
 import org.opentripplanner.graph_builder.module.osm.WayPropertySetSource.DrivingDirection;
 import org.opentripplanner.routing.api.request.RoutingRequest;
+import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.vertextype.IntersectionVertex;
 import org.slf4j.Logger;
